% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_mod.R
\name{mod_map_mod_ui}
\alias{mod_map_mod_ui}
\alias{mod_map_mod_server}
\title{mod_map_mod_ui and mod_map_mod_server}
\usage{
mod_map_mod_ui(id)

mod_map_mod_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
